--- src/main/java/io/supertokens/oauth/exceptions/OAuthAPIInvalidInputException.java
@@ -1,29 +1,0 @@
-/*
- *    Copyright (c) 2024, VRAI Labs and/or its affiliates. All rights reserved.
- *
- *    This software is licensed under the Apache License, Version 2.0 (the
- *    "License") as published by the Apache Software Foundation.
- *
- *    You may not use this file except in compliance with the License. You may
- *    obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- *    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- *    License for the specific language governing permissions and limitations
- *    under the License.
- */
-
-package io.supertokens.oauth.exceptions;
-
-import java.io.Serial;
-
-public class OAuthAPIInvalidInputException extends OAuthException{
-
-    @Serial
-    private static final long serialVersionUID = 665027786586190611L;
-
-    public OAuthAPIInvalidInputException(String error, String errorDescription) {
-        super(error, errorDescription, 400);
-    }
-}
