--- src/main/java/io/supertokens/oauth/exceptions/OAuthClientUpdateException.java
@@ -1,28 +1,0 @@
-/*
- *    Copyright (c) 2024, VRAI Labs and/or its affiliates. All rights reserved.
- *
- *    This software is licensed under the Apache License, Version 2.0 (the
- *    "License") as published by the Apache Software Foundation.
- *
- *    You may not use this file except in compliance with the License. You may
- *    obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- *    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- *    License for the specific language governing permissions and limitations
- *    under the License.
- */
-
-package io.supertokens.oauth.exceptions;
-
-import java.io.Serial;
-
-public class OAuthClientUpdateException extends OAuthException{
-    @Serial
-    private static final long serialVersionUID = -5191044905397936167L;
-
-    public OAuthClientUpdateException(String error, String errorDescription, int statusCode) {
-        super(error, errorDescription, statusCode);
-    }
-}
